refactor to create re-usable toggle component
