+++ package.json
@@ -19,17 +19,17 @@
     "@next/third-parties": "^14.2.7",
     "@tailwindcss/container-queries": "^0.1.1",
     "@tailwindcss/typography": "^0.5.15",
-    "@tanstack/react-query": "^5.52.2",
+    "@tanstack/react-query": "^5.53.1",
     "@types/node": "^22.5.1",
-    "@types/react": "^18.3.4",
+    "@types/react": "^18.3.5",
     "autoprefixer": "^10.4.20",
-    "axios": "^1.7.5",
+    "axios": "^1.7.6",
     "critters": "^0.0.24",
     "decanter": "^7.3.0",
     "graphql": "^16.9.0",
     "graphql-request": "^7.1.0",
     "graphql-tag": "^2.12.6",
-    "html-react-parser": "^5.1.14",
+    "html-react-parser": "^5.1.15",
     "jsona": "^1.12.1",
     "next": "^14.2.7",
     "next-drupal": "^1.6.0",
